ai-service/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
